requirements.txt
Brotli==1.0.9
rel==0.4.7
requests==2.27.1
websocket-client==1.3.1
